Special handling for Oem.Hp
The result for HP/HPE are different depending if the HTTP header
OData-Version is set or not. If OData-Version is _NOT_ set data are
returned in .Oem.Hp.links with endpoints in "href". If OData-Version
is set data are returned in .Oem.Hp.Links (note different case!) and
endpoints are defined as @odata.id. We always set "OData-Version: 4.0"